Create Newtonsoft.Json.13.0.3.nupkg

+++ EduzeniethFinal/EduzeniethFinal/Controllers/AdminController.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
+using System.Net;
 using System.Web;
 using System.Web.Mvc;
 
@@ -14,6 +15,10 @@ namespace EduzeniethFinal.Controllers
         // GET: Admin
         public ActionResult AddCourse()
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             return View();
         }
 
@@ -23,6 +28,10 @@ namespace EduzeniethFinal.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult AddCourse([Bind(Include = "Course_Code,Course_Name,teacherID,Course_desc")] Course course)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             if (ModelState.IsValid)
             {
                 course.Created_at = DateTime.Now; // Set Created_at here
@@ -36,12 +45,20 @@ namespace EduzeniethFinal.Controllers
 
         public ActionResult Home()
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             return View();
         }
 
         // GET: Admin/Manage_Students
         public ActionResult Manage_Students()
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var students = db.Students.ToList();
             return View(students);
         }
@@ -49,6 +66,10 @@ namespace EduzeniethFinal.Controllers
         // GET: Admin/Create
         public ActionResult Add_Student()
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             return View();
         }
 
@@ -57,6 +78,10 @@ namespace EduzeniethFinal.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Add_Student(Student student)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             if (ModelState.IsValid)
             {
                 db.Students.Add(student);
@@ -69,6 +94,10 @@ namespace EduzeniethFinal.Controllers
         // GET: Admin/Edit/5
         public ActionResult Edit(int id)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var student = db.Students.Find(id);
             if (student == null)
             {
@@ -82,6 +111,10 @@ namespace EduzeniethFinal.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(Student student)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(student).State = EntityState.Modified;
@@ -94,6 +127,10 @@ namespace EduzeniethFinal.Controllers
         // GET: Admin/Delete/5
         public ActionResult Delete_Student(int id)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var student = db.Students.Find(id);
             if (student == null)
             {
@@ -107,6 +144,10 @@ namespace EduzeniethFinal.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var student = db.Students.Find(id);
             db.Students.Remove(student);
             db.SaveChanges();
@@ -130,6 +171,10 @@ namespace EduzeniethFinal.Controllers
 
         public ActionResult Student_PendingRegistration()
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var pendingStudents = db.Students.Where(s => s.Status == 0).ToList();
             return View(pendingStudents);
         }
@@ -137,6 +182,10 @@ namespace EduzeniethFinal.Controllers
         [HttpPost]
         public ActionResult Student_PendingRegistration_Accept(int id)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var student = db.Students.Find(id);
             if (student != null)
             {
@@ -149,6 +198,10 @@ namespace EduzeniethFinal.Controllers
         [HttpPost]
         public ActionResult Student_PendingRegistration_Decline(int id)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var student = db.Students.Find(id);
             if (student != null)
             {
@@ -167,6 +220,10 @@ namespace EduzeniethFinal.Controllers
 
         public ActionResult TeacherPendingRegistration()
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var pendingTeachers = db.Teachers.Where(t => t.Status == 0).ToList();
             return View(pendingTeachers);
         }
@@ -175,6 +232,10 @@ namespace EduzeniethFinal.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult TeacherPendingRegistration_Accept(int id)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var teacher = db.Teachers.Find(id);
             if (teacher != null)
             {
@@ -188,6 +249,10 @@ namespace EduzeniethFinal.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult TeacherPendingRegistration_Decline(int id)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var teacher = db.Teachers.Find(id);
             if (teacher != null)
             {
@@ -217,12 +282,20 @@ namespace EduzeniethFinal.Controllers
 
         public ActionResult Student_Pending_Enrollment()
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var pendingStudents = db.Enrolls.Where(s => s.status == 0).ToList();
             return View(pendingStudents);
         }
         [HttpPost]
         public ActionResult Student_Pending_Enrollment_Accept(int id)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var student = db.Enrolls.Find(id);
             if (student != null)
             {
@@ -235,6 +308,10 @@ namespace EduzeniethFinal.Controllers
         [HttpPost]
         public ActionResult Student_Pending_Enrollment_Decline(int id)
         {
+            if (Session["admin"] == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             var student = db.Enrolls.Find(id);
             if (student != null)
             {
@@ -244,6 +321,121 @@ namespace EduzeniethFinal.Controllers
             return RedirectToAction("Student_Pending_Enrollment");
         }
 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+        public ActionResult Login()
+        {
+            return View();
+        }
+
+        // POST: Student/Login
+        [HttpPost]
+        public ActionResult Login(string username, string password)
+        {
+
+            if (username.Equals("admin") && password.Equals("123456"))
+            {
+                // Login successful
+                // Redirect to dashboard or any other page
+                Session["admin"] = "logged_in";
+                return RedirectToAction("Home", "Admin");
+            }
+            else
+            {
+                ViewBag.Log = "Login Failed.Please Enter Correct Username and Password";
+                // Login failed
+                // You can add a ViewBag message here to show an error message on the login page
+                ViewBag.ErrorMessage = "Invalid username or password";
+                return View();
+            }
+
+        }
+        public ActionResult Logout()
+        {
+            Session["admin"] = null;
+            return RedirectToAction("Index", "Home");
+        }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+        public ActionResult Manage_Courses()
+        {
+            return View(db.Courses.ToList());
+        }
+        // GET: Courses/Delete/5
+        public ActionResult Delete_Course(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            Course course = db.Courses.Find(id);
+            if (course == null)
+            {
+                return HttpNotFound();
+            }
+            return View(course);
+        }
+
+        // POST: Courses/Delete/5
+        [HttpPost, ActionName("Delete_Course")]
+        [ValidateAntiForgeryToken]
+        public ActionResult DeleteConfirmed_Course(int id)
+        {
+            Course course = db.Courses.Find(id);
+            db.Courses.Remove(course);
+            db.SaveChanges();
+            return RedirectToAction("Delete_Course","Courses");
+        }
